handle backTo param
