+++ src/libs/Navigation/linkingConfig/FULL_SCREEN_TO_RHP_MAPPING.ts
@@ -1,5 +1,5 @@
+import type {FullScreenName} from '@libs/Navigation/types';
 import SCREENS from '@src/SCREENS';
-import type {FullScreenName} from './types';
 
 const FULL_SCREEN_TO_RHP_MAPPING: Partial<Record<FullScreenName, string[]>> = {
     [SCREENS.SETTINGS.PROFILE.ROOT]: [
