+++ src/libs/Navigation/linkingConfig/getMatchingBottomTabRouteForState.ts
@@ -1,8 +1,7 @@
-// import CONST from '@src/CONST';
+import getTopmostCentralPaneRoute from '@libs/Navigation/getTopmostCentralPaneRoute';
+import type {BottomTabName, NavigationPartialRoute, RootStackParamList, State} from '@libs/Navigation/types';
 import SCREENS from '@src/SCREENS';
-import getTopmostCentralPaneRoute from './getTopmostCentralPaneRoute';
 import {CENTRAL_PANE_TO_TAB_MAPPING} from './TAB_TO_CENTRAL_PANE_MAPPING';
-import type {BottomTabName, NavigationPartialRoute, RootStackParamList, State} from './types';
 
 // Get the route that matches the topmost central pane route in the navigation stack. e.g REPORT -> HOME
 function getMatchingBottomTabRouteForState(state: State<RootStackParamList>): NavigationPartialRoute<BottomTabName> {
